修正docker启动脚本

+++ start.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
-rm -rf /xuexi/code/TechXueXi
-git -C /xuexi/code/TechXueXi config pull.ff only
+if ! git -C /xuexi/code/TechXueXi config pull.ff only; then
+    rm -rf /xuexi/code/TechXueXi
+    cd /xuexi/code/ && git clone -b ${pullbranche} ${Sourcepath}
+fi
 printenv >> /etc/environment
 touch /var/log/cron.log
 ./run.sh 2>&1 &
